app/requirements.txt
boto3==1.26.34
configparser==5.3.0
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-graphql-jwt==0.3.4
django-ipware==4.0.2
django-redis==5.2.0
django-safedelete==1.3.1
django-storages==1.13.1
django-tenants==3.4.7
djangorestframework==3.14.0
graphene-django==3.0.0
gunicorn==20.1.0
pre-commit==2.21.0
psycopg2-binary==2.9.5
redis-py-cluster==2.1.3
ssm-parameter-store==19.11.0
validators==0.20.0
